Split shell computer control tests
